url open to check the sttus
